+++ xknx/remote_value/remote_value_color_xyy.py
@@ -14,4 +14,5 @@ from .remote_value import RemoteValue
 class RemoteValueColorXYY(RemoteValue[XYYColor]):
     """Abstraction for remote value of KNX DPT 242.600 (DPT_Colour_xyY)."""
 
+    __slots__ = ()
     dpt_class = DPTColorXYY